Update the detail() view and vote() view.
Making change on polls/views.py file. I've carefully
read line by line in the tutorial to understand the new
stuffs on it, however, it's still something I've to look
up on Google to figure it out the what are these build-in
functions do, and how does it works on the program.